pyproject.toml
[tool.poetry]
name = "astronomy-database"
version = "0.1.0"
description = ""
authors = ["manickaug <manickaug@vcu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
matplotlib = "^3.8.0"
plotly = "^5.18.0"
jupyter = "^1.0.0"
tabulate = "^0.9.0"
markdown = "^3.5"
pandas = "^2.1.1"
sqlalchemy = "^2.0.22"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"
astronomy-engine = "^2.1.17"
pymysql = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
